Decouple ScoreHolder from ScoreDefinition
