+++ optaplanner-core/src/main/java/org/optaplanner/core/impl/score/buildin/bendablebigdecimal/BendableBigDecimalScoreDefinition.java
@@ -112,11 +112,6 @@ public class BendableBigDecimalScoreDefinition extends AbstractBendableScoreDefi
     }
 
     @Override
-    public BendableBigDecimalScoreHolderImpl buildScoreHolder(boolean constraintMatchEnabled) {
-        return new BendableBigDecimalScoreHolderImpl(constraintMatchEnabled, hardLevelsSize, softLevelsSize);
-    }
-
-    @Override
     public BendableBigDecimalScore buildOptimisticBound(InitializingScoreTrend initializingScoreTrend,
             BendableBigDecimalScore score) {
         // TODO https://issues.redhat.com/browse/PLANNER-232
