+++ optaplanner-core/src/main/java/org/optaplanner/core/impl/score/director/drools/holder/BendableBigDecimalScoreHolderImpl.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.optaplanner.core.impl.score.buildin.bendablelong;
+package org.optaplanner.core.impl.score.director.drools.holder;
 
 import java.math.BigDecimal;
 import java.util.Arrays;
@@ -26,27 +26,29 @@ import org.kie.api.runtime.rule.RuleContext;
 import org.optaplanner.core.api.domain.constraintweight.ConstraintConfiguration;
 import org.optaplanner.core.api.domain.constraintweight.ConstraintWeight;
 import org.optaplanner.core.api.domain.solution.PlanningScore;
-import org.optaplanner.core.api.score.buildin.bendablelong.BendableLongScore;
-import org.optaplanner.core.api.score.buildin.bendablelong.BendableLongScoreHolder;
+import org.optaplanner.core.api.score.buildin.bendablebigdecimal.BendableBigDecimalScore;
+import org.optaplanner.core.api.score.buildin.bendablebigdecimal.BendableBigDecimalScoreHolder;
 import org.optaplanner.core.impl.score.holder.AbstractScoreHolder;
 
 /**
- * @see BendableLongScore
+ * @see BendableBigDecimalScore
  */
-public final class BendableLongScoreHolderImpl extends AbstractScoreHolder<BendableLongScore>
-        implements BendableLongScoreHolder {
+public final class BendableBigDecimalScoreHolderImpl extends AbstractScoreHolder<BendableBigDecimalScore>
+        implements BendableBigDecimalScoreHolder {
 
-    protected final Map<Rule, LongMatchExecutor> matchExecutorByNumberMap = new LinkedHashMap<>();
+    protected final Map<Rule, BigDecimalMatchExecutor> matchExecutorByNumberMap = new LinkedHashMap<>();
     /** Slower than {@link #matchExecutorByNumberMap} */
-    protected final Map<Rule, ScoreMatchExecutor<BendableLongScore>> matchExecutorByScoreMap = new LinkedHashMap<>();
+    protected final Map<Rule, ScoreMatchExecutor<BendableBigDecimalScore>> matchExecutorByScoreMap = new LinkedHashMap<>();
 
-    private long[] hardScores;
-    private long[] softScores;
+    private BigDecimal[] hardScores;
+    private BigDecimal[] softScores;
 
-    public BendableLongScoreHolderImpl(boolean constraintMatchEnabled, int hardLevelsSize, int softLevelsSize) {
+    public BendableBigDecimalScoreHolderImpl(boolean constraintMatchEnabled, int hardLevelsSize, int softLevelsSize) {
         super(constraintMatchEnabled);
-        hardScores = new long[hardLevelsSize];
-        softScores = new long[softLevelsSize];
+        hardScores = new BigDecimal[hardLevelsSize];
+        Arrays.fill(hardScores, BigDecimal.ZERO);
+        softScores = new BigDecimal[softLevelsSize];
+        Arrays.fill(softScores, BigDecimal.ZERO);
     }
 
     @Override
@@ -54,7 +56,7 @@ public final class BendableLongScoreHolderImpl extends AbstractScoreHolder<Benda
         return hardScores.length;
     }
 
-    public long getHardScore(int hardLevel) {
+    public BigDecimal getHardScore(int hardLevel) {
         return hardScores[hardLevel];
     }
 
@@ -63,7 +65,7 @@ public final class BendableLongScoreHolderImpl extends AbstractScoreHolder<Benda
         return softScores.length;
     }
 
-    public long getSoftScore(int softLevel) {
+    public BigDecimal getSoftScore(int softLevel) {
         return softScores[softLevel];
     }
 
@@ -72,18 +74,18 @@ public final class BendableLongScoreHolderImpl extends AbstractScoreHolder<Benda
     // ************************************************************************
 
     @Override
-    public void configureConstraintWeight(Rule rule, BendableLongScore constraintWeight) {
+    public void configureConstraintWeight(Rule rule, BendableBigDecimalScore constraintWeight) {
         super.configureConstraintWeight(rule, constraintWeight);
-        LongMatchExecutor matchExecutor;
-        if (constraintWeight.equals(BendableLongScore.zero(hardScores.length, softScores.length))) {
-            matchExecutor = (RuleContext kcontext, long matchWeight) -> {
+        BigDecimalMatchExecutor matchExecutor;
+        if (constraintWeight.equals(BendableBigDecimalScore.zero(hardScores.length, softScores.length))) {
+            matchExecutor = (RuleContext kcontext, BigDecimal matchWeight) -> {
             };
         } else {
             Integer singleLevel = null;
-            Long singleLevelWeight = null;
+            BigDecimal singleLevelWeight = null;
             for (int i = 0; i < constraintWeight.getLevelsSize(); i++) {
-                long levelWeight = constraintWeight.getHardOrSoftScore(i);
-                if (levelWeight != 0L) {
+                BigDecimal levelWeight = constraintWeight.getHardOrSoftScore(i);
+                if (!levelWeight.equals(BigDecimal.ZERO)) {
                     if (singleLevel != null) {
                         singleLevel = null;
                         singleLevelWeight = null;
@@ -94,41 +96,41 @@ public final class BendableLongScoreHolderImpl extends AbstractScoreHolder<Benda
                 }
             }
             if (singleLevel != null) {
-                long levelWeight = singleLevelWeight;
+                BigDecimal levelWeight = singleLevelWeight;
                 if (singleLevel < constraintWeight.getHardLevelsSize()) {
                     int level = singleLevel;
                     matchExecutor =
-                            (RuleContext kcontext, long matchWeight) -> addHardConstraintMatch(
-                                    kcontext, level, levelWeight * matchWeight);
+                            (RuleContext kcontext, BigDecimal matchWeight) -> addHardConstraintMatch(
+                                    kcontext, level, levelWeight.multiply(matchWeight));
                 } else {
                     int level = singleLevel - constraintWeight.getHardLevelsSize();
                     matchExecutor =
-                            (RuleContext kcontext, long matchWeight) -> addSoftConstraintMatch(
-                                    kcontext, level, levelWeight * matchWeight);
+                            (RuleContext kcontext, BigDecimal matchWeight) -> addSoftConstraintMatch(
+                                    kcontext, level, levelWeight.multiply(matchWeight));
                 }
             } else {
-                matchExecutor = (RuleContext kcontext, long matchWeight) -> {
-                    long[] hardWeights = new long[hardScores.length];
-                    long[] softWeights = new long[softScores.length];
+                matchExecutor = (RuleContext kcontext, BigDecimal matchWeight) -> {
+                    BigDecimal[] hardWeights = new BigDecimal[hardScores.length];
+                    BigDecimal[] softWeights = new BigDecimal[softScores.length];
                     for (int i = 0; i < hardWeights.length; i++) {
-                        hardWeights[i] = constraintWeight.getHardScore(i) * matchWeight;
+                        hardWeights[i] = constraintWeight.getHardScore(i).multiply(matchWeight);
                     }
                     for (int i = 0; i < softWeights.length; i++) {
-                        softWeights[i] = constraintWeight.getSoftScore(i) * matchWeight;
+                        softWeights[i] = constraintWeight.getSoftScore(i).multiply(matchWeight);
                     }
                     addMultiConstraintMatch(kcontext, hardWeights, softWeights);
                 };
             }
         }
         matchExecutorByNumberMap.put(rule, matchExecutor);
-        matchExecutorByScoreMap.put(rule, (RuleContext kcontext, BendableLongScore weightMultiplier) -> {
-            long[] hardWeights = new long[hardScores.length];
-            long[] softWeights = new long[softScores.length];
+        matchExecutorByScoreMap.put(rule, (RuleContext kcontext, BendableBigDecimalScore weightMultiplier) -> {
+            BigDecimal[] hardWeights = new BigDecimal[hardScores.length];
+            BigDecimal[] softWeights = new BigDecimal[softScores.length];
             for (int i = 0; i < hardWeights.length; i++) {
-                hardWeights[i] = constraintWeight.getHardScore(i) * weightMultiplier.getHardScore(i);
+                hardWeights[i] = constraintWeight.getHardScore(i).multiply(weightMultiplier.getHardScore(i));
             }
             for (int i = 0; i < softWeights.length; i++) {
-                softWeights[i] = constraintWeight.getSoftScore(i) * weightMultiplier.getSoftScore(i);
+                softWeights[i] = constraintWeight.getSoftScore(i).multiply(weightMultiplier.getSoftScore(i));
             }
             addMultiConstraintMatch(kcontext, hardWeights, softWeights);
         });
@@ -140,56 +142,61 @@ public final class BendableLongScoreHolderImpl extends AbstractScoreHolder<Benda
 
     @Override
     public void penalize(RuleContext kcontext) {
-        impactScore(kcontext, -1L);
+        impactScore(kcontext, BigDecimal.ONE.negate());
     }
 
     @Override
-    public void penalize(RuleContext kcontext, long weightMultiplier) {
-        impactScore(kcontext, -weightMultiplier);
+    public void penalize(RuleContext kcontext, BigDecimal weightMultiplier) {
+        impactScore(kcontext, weightMultiplier.negate());
     }
 
     @Override
-    public void penalize(RuleContext kcontext, long[] hardWeightsMultiplier, long[] softWeightsMultiplier) {
-        long[] negatedHardWeightsMultiplier = new long[hardScores.length];
-        long[] negatedSoftWeightsMultiplier = new long[softScores.length];
+    public void penalize(RuleContext kcontext, BigDecimal[] hardWeightsMultiplier, BigDecimal[] softWeightsMultiplier) {
+        BigDecimal[] negatedHardWeightsMultiplier = new BigDecimal[hardScores.length];
+        BigDecimal[] negatedSoftWeightsMultiplier = new BigDecimal[softScores.length];
         for (int i = 0; i < negatedHardWeightsMultiplier.length; i++) {
-            negatedHardWeightsMultiplier[i] = -hardWeightsMultiplier[i];
+            negatedHardWeightsMultiplier[i] = hardWeightsMultiplier[i].negate();
         }
         for (int i = 0; i < negatedSoftWeightsMultiplier.length; i++) {
-            negatedSoftWeightsMultiplier[i] = -softWeightsMultiplier[i];
+            negatedSoftWeightsMultiplier[i] = softWeightsMultiplier[i].negate();
         }
         impactScore(kcontext, negatedHardWeightsMultiplier, negatedSoftWeightsMultiplier);
     }
 
     @Override
     public void reward(RuleContext kcontext) {
-        impactScore(kcontext, 1L);
+        impactScore(kcontext, BigDecimal.ONE);
     }
 
     @Override
-    public void reward(RuleContext kcontext, long weightMultiplier) {
+    public void reward(RuleContext kcontext, BigDecimal weightMultiplier) {
         impactScore(kcontext, weightMultiplier);
     }
 
     @Override
-    public void reward(RuleContext kcontext, long[] hardWeightsMultiplier, long[] softWeightsMultiplier) {
+    public void reward(RuleContext kcontext, BigDecimal[] hardWeightsMultiplier, BigDecimal[] softWeightsMultiplier) {
         impactScore(kcontext, hardWeightsMultiplier, softWeightsMultiplier);
     }
 
     @Override
     public void impactScore(RuleContext kcontext) {
-        impactScore(kcontext, 1L);
+        impactScore(kcontext, BigDecimal.ONE);
     }
 
     @Override
     public void impactScore(RuleContext kcontext, int weightMultiplier) {
-        impactScore(kcontext, (long) weightMultiplier);
+        impactScore(kcontext, BigDecimal.valueOf(weightMultiplier));
     }
 
     @Override
     public void impactScore(RuleContext kcontext, long weightMultiplier) {
+        impactScore(kcontext, BigDecimal.valueOf(weightMultiplier));
+    }
+
+    @Override
+    public void impactScore(RuleContext kcontext, BigDecimal weightMultiplier) {
         Rule rule = kcontext.getRule();
-        LongMatchExecutor matchExecutor = matchExecutorByNumberMap.get(rule);
+        BigDecimalMatchExecutor matchExecutor = matchExecutorByNumberMap.get(rule);
         if (matchExecutor == null) {
             throw new IllegalStateException("The DRL rule (" + rule.getPackageName() + ":" + rule.getName()
                     + ") does not match a @" + ConstraintWeight.class.getSimpleName() + " on the @"
@@ -198,23 +205,15 @@ public final class BendableLongScoreHolderImpl extends AbstractScoreHolder<Benda
         matchExecutor.accept(kcontext, weightMultiplier);
     }
 
-    @Override
-    public void impactScore(RuleContext kcontext, BigDecimal weightMultiplier) {
-        throw new UnsupportedOperationException("In the rule (" + kcontext.getRule().getName()
-                + "), the scoreHolder class (" + getClass()
-                + ") does not support a BigDecimal weightMultiplier (" + weightMultiplier + ").\n"
-                + "If you're using constraint streams, maybe switch from penalizeBigDecimal() to penalizeLong().");
-    }
-
-    private void impactScore(RuleContext kcontext, long[] hardWeightsMultiplier, long[] softWeightsMultiplier) {
+    private void impactScore(RuleContext kcontext, BigDecimal[] hardWeightsMultiplier, BigDecimal[] softWeightsMultiplier) {
         Rule rule = kcontext.getRule();
-        ScoreMatchExecutor<BendableLongScore> matchExecutor = matchExecutorByScoreMap.get(rule);
+        ScoreMatchExecutor<BendableBigDecimalScore> matchExecutor = matchExecutorByScoreMap.get(rule);
         if (matchExecutor == null) {
             throw new IllegalStateException("The DRL rule (" + rule.getPackageName() + ":" + rule.getName()
                     + ") does not match a @" + ConstraintWeight.class.getSimpleName() + " on the @"
                     + ConstraintConfiguration.class.getSimpleName() + " annotated class.");
         }
-        matchExecutor.accept(kcontext, BendableLongScore.of(hardWeightsMultiplier, softWeightsMultiplier));
+        matchExecutor.accept(kcontext, BendableBigDecimalScore.of(hardWeightsMultiplier, softWeightsMultiplier));
     }
 
     // ************************************************************************
@@ -222,50 +221,54 @@ public final class BendableLongScoreHolderImpl extends AbstractScoreHolder<Benda
     // ************************************************************************
 
     @Override
-    public void addHardConstraintMatch(RuleContext kcontext, int hardLevel, long weight) {
+    public void addHardConstraintMatch(RuleContext kcontext, int hardLevel, BigDecimal weight) {
         if (hardLevel >= hardScores.length) {
             throw new IllegalArgumentException("The hardLevel (" + hardLevel
                     + ") isn't lower than the hardScores length (" + hardScores.length
                     + ") defined by the @" + PlanningScore.class.getSimpleName() + " on the planning solution class.");
         }
-        hardScores[hardLevel] += weight;
+        hardScores[hardLevel] = hardScores[hardLevel].add(weight);
         registerConstraintMatch(kcontext,
-                () -> hardScores[hardLevel] -= weight,
+                () -> hardScores[hardLevel] = hardScores[hardLevel].subtract(weight),
                 () -> {
-                    long[] newHardScores = new long[hardScores.length];
-                    long[] newSoftScores = new long[softScores.length];
+                    BigDecimal[] newHardScores = new BigDecimal[hardScores.length];
+                    Arrays.fill(newHardScores, BigDecimal.ZERO);
+                    BigDecimal[] newSoftScores = new BigDecimal[softScores.length];
+                    Arrays.fill(newSoftScores, BigDecimal.ZERO);
                     newHardScores[hardLevel] = weight;
-                    return BendableLongScore.of(newHardScores, newSoftScores);
+                    return BendableBigDecimalScore.of(newHardScores, newSoftScores);
                 });
     }
 
     @Override
-    public void addSoftConstraintMatch(RuleContext kcontext, int softLevel, long weight) {
+    public void addSoftConstraintMatch(RuleContext kcontext, int softLevel, BigDecimal weight) {
         if (softLevel >= softScores.length) {
             throw new IllegalArgumentException("The softLevel (" + softLevel
                     + ") isn't lower than the softScores length (" + softScores.length
                     + ") defined by the @" + PlanningScore.class.getSimpleName() + " on the planning solution class.");
         }
-        softScores[softLevel] += weight;
+        softScores[softLevel] = softScores[softLevel].add(weight);
         registerConstraintMatch(kcontext,
-                () -> softScores[softLevel] -= weight,
+                () -> softScores[softLevel] = softScores[softLevel].subtract(weight),
                 () -> {
-                    long[] newHardScores = new long[hardScores.length];
-                    long[] newSoftScores = new long[softScores.length];
+                    BigDecimal[] newHardScores = new BigDecimal[hardScores.length];
+                    Arrays.fill(newHardScores, BigDecimal.ZERO);
+                    BigDecimal[] newSoftScores = new BigDecimal[softScores.length];
+                    Arrays.fill(newSoftScores, BigDecimal.ZERO);
                     newSoftScores[softLevel] = weight;
-                    return BendableLongScore.of(newHardScores, newSoftScores);
+                    return BendableBigDecimalScore.of(newHardScores, newSoftScores);
                 });
     }
 
     @Override
-    public void addMultiConstraintMatch(RuleContext kcontext, long[] hardWeights, long[] softWeights) {
+    public void addMultiConstraintMatch(RuleContext kcontext, BigDecimal[] hardWeights, BigDecimal[] softWeights) {
         if (hardWeights.length != hardScores.length) {
             throw new IllegalArgumentException("The hardWeights length (" + hardWeights.length
                     + ") is different than the hardScores length (" + hardScores.length
                     + ") defined by the @" + PlanningScore.class.getSimpleName() + " on the planning solution class.");
         }
         for (int i = 0; i < hardScores.length; i++) {
-            hardScores[i] += hardWeights[i];
+            hardScores[i] = hardScores[i].add(hardWeights[i]);
         }
         if (softWeights.length != softScores.length) {
             throw new IllegalArgumentException("The softWeights length (" + softWeights.length
@@ -273,23 +276,23 @@ public final class BendableLongScoreHolderImpl extends AbstractScoreHolder<Benda
                     + ") defined by the @" + PlanningScore.class.getSimpleName() + " on the planning solution class.");
         }
         for (int i = 0; i < softScores.length; i++) {
-            softScores[i] += softWeights[i];
+            softScores[i] = softScores[i].add(softWeights[i]);
         }
         registerConstraintMatch(kcontext,
                 () -> {
                     for (int i = 0; i < hardScores.length; i++) {
-                        hardScores[i] -= hardWeights[i];
+                        hardScores[i] = hardScores[i].subtract(hardWeights[i]);
                     }
                     for (int i = 0; i < softScores.length; i++) {
-                        softScores[i] -= softWeights[i];
+                        softScores[i] = softScores[i].subtract(softWeights[i]);
                     }
                 },
-                () -> BendableLongScore.of(hardWeights, softWeights));
+                () -> BendableBigDecimalScore.of(hardWeights, softWeights));
     }
 
     @Override
-    public BendableLongScore extractScore(int initScore) {
-        return BendableLongScore.ofUninitialized(initScore,
+    public BendableBigDecimalScore extractScore(int initScore) {
+        return BendableBigDecimalScore.ofUninitialized(initScore,
                 Arrays.copyOf(hardScores, hardScores.length),
                 Arrays.copyOf(softScores, softScores.length));
     }
