+++ optaplanner-core/src/main/java/org/optaplanner/core/impl/score/director/drools/holder/BendableLongScoreHolderImpl.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.optaplanner.core.impl.score.buildin.bendablebigdecimal;
+package org.optaplanner.core.impl.score.director.drools.holder;
 
 import java.math.BigDecimal;
 import java.util.Arrays;
@@ -26,29 +26,27 @@ import org.kie.api.runtime.rule.RuleContext;
 import org.optaplanner.core.api.domain.constraintweight.ConstraintConfiguration;
 import org.optaplanner.core.api.domain.constraintweight.ConstraintWeight;
 import org.optaplanner.core.api.domain.solution.PlanningScore;
-import org.optaplanner.core.api.score.buildin.bendablebigdecimal.BendableBigDecimalScore;
-import org.optaplanner.core.api.score.buildin.bendablebigdecimal.BendableBigDecimalScoreHolder;
+import org.optaplanner.core.api.score.buildin.bendablelong.BendableLongScore;
+import org.optaplanner.core.api.score.buildin.bendablelong.BendableLongScoreHolder;
 import org.optaplanner.core.impl.score.holder.AbstractScoreHolder;
 
 /**
- * @see BendableBigDecimalScore
+ * @see BendableLongScore
  */
-public final class BendableBigDecimalScoreHolderImpl extends AbstractScoreHolder<BendableBigDecimalScore>
-        implements BendableBigDecimalScoreHolder {
+public final class BendableLongScoreHolderImpl extends AbstractScoreHolder<BendableLongScore>
+        implements BendableLongScoreHolder {
 
-    protected final Map<Rule, BigDecimalMatchExecutor> matchExecutorByNumberMap = new LinkedHashMap<>();
+    protected final Map<Rule, LongMatchExecutor> matchExecutorByNumberMap = new LinkedHashMap<>();
     /** Slower than {@link #matchExecutorByNumberMap} */
-    protected final Map<Rule, ScoreMatchExecutor<BendableBigDecimalScore>> matchExecutorByScoreMap = new LinkedHashMap<>();
+    protected final Map<Rule, ScoreMatchExecutor<BendableLongScore>> matchExecutorByScoreMap = new LinkedHashMap<>();
 
-    private BigDecimal[] hardScores;
-    private BigDecimal[] softScores;
+    private long[] hardScores;
+    private long[] softScores;
 
-    public BendableBigDecimalScoreHolderImpl(boolean constraintMatchEnabled, int hardLevelsSize, int softLevelsSize) {
+    public BendableLongScoreHolderImpl(boolean constraintMatchEnabled, int hardLevelsSize, int softLevelsSize) {
         super(constraintMatchEnabled);
-        hardScores = new BigDecimal[hardLevelsSize];
-        Arrays.fill(hardScores, BigDecimal.ZERO);
-        softScores = new BigDecimal[softLevelsSize];
-        Arrays.fill(softScores, BigDecimal.ZERO);
+        hardScores = new long[hardLevelsSize];
+        softScores = new long[softLevelsSize];
     }
 
     @Override
@@ -56,7 +54,7 @@ public final class BendableBigDecimalScoreHolderImpl extends AbstractScoreHolder
         return hardScores.length;
     }
 
-    public BigDecimal getHardScore(int hardLevel) {
+    public long getHardScore(int hardLevel) {
         return hardScores[hardLevel];
     }
 
@@ -65,7 +63,7 @@ public final class BendableBigDecimalScoreHolderImpl extends AbstractScoreHolder
         return softScores.length;
     }
 
-    public BigDecimal getSoftScore(int softLevel) {
+    public long getSoftScore(int softLevel) {
         return softScores[softLevel];
     }
 
@@ -74,18 +72,18 @@ public final class BendableBigDecimalScoreHolderImpl extends AbstractScoreHolder
     // ************************************************************************
 
     @Override
-    public void configureConstraintWeight(Rule rule, BendableBigDecimalScore constraintWeight) {
+    public void configureConstraintWeight(Rule rule, BendableLongScore constraintWeight) {
         super.configureConstraintWeight(rule, constraintWeight);
-        BigDecimalMatchExecutor matchExecutor;
-        if (constraintWeight.equals(BendableBigDecimalScore.zero(hardScores.length, softScores.length))) {
-            matchExecutor = (RuleContext kcontext, BigDecimal matchWeight) -> {
+        LongMatchExecutor matchExecutor;
+        if (constraintWeight.equals(BendableLongScore.zero(hardScores.length, softScores.length))) {
+            matchExecutor = (RuleContext kcontext, long matchWeight) -> {
             };
         } else {
             Integer singleLevel = null;
-            BigDecimal singleLevelWeight = null;
+            Long singleLevelWeight = null;
             for (int i = 0; i < constraintWeight.getLevelsSize(); i++) {
-                BigDecimal levelWeight = constraintWeight.getHardOrSoftScore(i);
-                if (!levelWeight.equals(BigDecimal.ZERO)) {
+                long levelWeight = constraintWeight.getHardOrSoftScore(i);
+                if (levelWeight != 0L) {
                     if (singleLevel != null) {
                         singleLevel = null;
                         singleLevelWeight = null;
@@ -96,41 +94,41 @@ public final class BendableBigDecimalScoreHolderImpl extends AbstractScoreHolder
                 }
             }
             if (singleLevel != null) {
-                BigDecimal levelWeight = singleLevelWeight;
+                long levelWeight = singleLevelWeight;
                 if (singleLevel < constraintWeight.getHardLevelsSize()) {
                     int level = singleLevel;
                     matchExecutor =
-                            (RuleContext kcontext, BigDecimal matchWeight) -> addHardConstraintMatch(
-                                    kcontext, level, levelWeight.multiply(matchWeight));
+                            (RuleContext kcontext, long matchWeight) -> addHardConstraintMatch(
+                                    kcontext, level, levelWeight * matchWeight);
                 } else {
                     int level = singleLevel - constraintWeight.getHardLevelsSize();
                     matchExecutor =
-                            (RuleContext kcontext, BigDecimal matchWeight) -> addSoftConstraintMatch(
-                                    kcontext, level, levelWeight.multiply(matchWeight));
+                            (RuleContext kcontext, long matchWeight) -> addSoftConstraintMatch(
+                                    kcontext, level, levelWeight * matchWeight);
                 }
             } else {
-                matchExecutor = (RuleContext kcontext, BigDecimal matchWeight) -> {
-                    BigDecimal[] hardWeights = new BigDecimal[hardScores.length];
-                    BigDecimal[] softWeights = new BigDecimal[softScores.length];
+                matchExecutor = (RuleContext kcontext, long matchWeight) -> {
+                    long[] hardWeights = new long[hardScores.length];
+                    long[] softWeights = new long[softScores.length];
                     for (int i = 0; i < hardWeights.length; i++) {
-                        hardWeights[i] = constraintWeight.getHardScore(i).multiply(matchWeight);
+                        hardWeights[i] = constraintWeight.getHardScore(i) * matchWeight;
                     }
                     for (int i = 0; i < softWeights.length; i++) {
-                        softWeights[i] = constraintWeight.getSoftScore(i).multiply(matchWeight);
+                        softWeights[i] = constraintWeight.getSoftScore(i) * matchWeight;
                     }
                     addMultiConstraintMatch(kcontext, hardWeights, softWeights);
                 };
             }
         }
         matchExecutorByNumberMap.put(rule, matchExecutor);
-        matchExecutorByScoreMap.put(rule, (RuleContext kcontext, BendableBigDecimalScore weightMultiplier) -> {
-            BigDecimal[] hardWeights = new BigDecimal[hardScores.length];
-            BigDecimal[] softWeights = new BigDecimal[softScores.length];
+        matchExecutorByScoreMap.put(rule, (RuleContext kcontext, BendableLongScore weightMultiplier) -> {
+            long[] hardWeights = new long[hardScores.length];
+            long[] softWeights = new long[softScores.length];
             for (int i = 0; i < hardWeights.length; i++) {
-                hardWeights[i] = constraintWeight.getHardScore(i).multiply(weightMultiplier.getHardScore(i));
+                hardWeights[i] = constraintWeight.getHardScore(i) * weightMultiplier.getHardScore(i);
             }
             for (int i = 0; i < softWeights.length; i++) {
-                softWeights[i] = constraintWeight.getSoftScore(i).multiply(weightMultiplier.getSoftScore(i));
+                softWeights[i] = constraintWeight.getSoftScore(i) * weightMultiplier.getSoftScore(i);
             }
             addMultiConstraintMatch(kcontext, hardWeights, softWeights);
         });
@@ -142,61 +140,56 @@ public final class BendableBigDecimalScoreHolderImpl extends AbstractScoreHolder
 
     @Override
     public void penalize(RuleContext kcontext) {
-        impactScore(kcontext, BigDecimal.ONE.negate());
+        impactScore(kcontext, -1L);
     }
 
     @Override
-    public void penalize(RuleContext kcontext, BigDecimal weightMultiplier) {
-        impactScore(kcontext, weightMultiplier.negate());
+    public void penalize(RuleContext kcontext, long weightMultiplier) {
+        impactScore(kcontext, -weightMultiplier);
     }
 
     @Override
-    public void penalize(RuleContext kcontext, BigDecimal[] hardWeightsMultiplier, BigDecimal[] softWeightsMultiplier) {
-        BigDecimal[] negatedHardWeightsMultiplier = new BigDecimal[hardScores.length];
-        BigDecimal[] negatedSoftWeightsMultiplier = new BigDecimal[softScores.length];
+    public void penalize(RuleContext kcontext, long[] hardWeightsMultiplier, long[] softWeightsMultiplier) {
+        long[] negatedHardWeightsMultiplier = new long[hardScores.length];
+        long[] negatedSoftWeightsMultiplier = new long[softScores.length];
         for (int i = 0; i < negatedHardWeightsMultiplier.length; i++) {
-            negatedHardWeightsMultiplier[i] = hardWeightsMultiplier[i].negate();
+            negatedHardWeightsMultiplier[i] = -hardWeightsMultiplier[i];
         }
         for (int i = 0; i < negatedSoftWeightsMultiplier.length; i++) {
-            negatedSoftWeightsMultiplier[i] = softWeightsMultiplier[i].negate();
+            negatedSoftWeightsMultiplier[i] = -softWeightsMultiplier[i];
         }
         impactScore(kcontext, negatedHardWeightsMultiplier, negatedSoftWeightsMultiplier);
     }
 
     @Override
     public void reward(RuleContext kcontext) {
-        impactScore(kcontext, BigDecimal.ONE);
+        impactScore(kcontext, 1L);
     }
 
     @Override
-    public void reward(RuleContext kcontext, BigDecimal weightMultiplier) {
+    public void reward(RuleContext kcontext, long weightMultiplier) {
         impactScore(kcontext, weightMultiplier);
     }
 
     @Override
-    public void reward(RuleContext kcontext, BigDecimal[] hardWeightsMultiplier, BigDecimal[] softWeightsMultiplier) {
+    public void reward(RuleContext kcontext, long[] hardWeightsMultiplier, long[] softWeightsMultiplier) {
         impactScore(kcontext, hardWeightsMultiplier, softWeightsMultiplier);
     }
 
     @Override
     public void impactScore(RuleContext kcontext) {
-        impactScore(kcontext, BigDecimal.ONE);
+        impactScore(kcontext, 1L);
     }
 
     @Override
     public void impactScore(RuleContext kcontext, int weightMultiplier) {
-        impactScore(kcontext, BigDecimal.valueOf(weightMultiplier));
+        impactScore(kcontext, (long) weightMultiplier);
     }
 
     @Override
     public void impactScore(RuleContext kcontext, long weightMultiplier) {
-        impactScore(kcontext, BigDecimal.valueOf(weightMultiplier));
-    }
-
-    @Override
-    public void impactScore(RuleContext kcontext, BigDecimal weightMultiplier) {
         Rule rule = kcontext.getRule();
-        BigDecimalMatchExecutor matchExecutor = matchExecutorByNumberMap.get(rule);
+        LongMatchExecutor matchExecutor = matchExecutorByNumberMap.get(rule);
         if (matchExecutor == null) {
             throw new IllegalStateException("The DRL rule (" + rule.getPackageName() + ":" + rule.getName()
                     + ") does not match a @" + ConstraintWeight.class.getSimpleName() + " on the @"
@@ -205,15 +198,23 @@ public final class BendableBigDecimalScoreHolderImpl extends AbstractScoreHolder
         matchExecutor.accept(kcontext, weightMultiplier);
     }
 
-    private void impactScore(RuleContext kcontext, BigDecimal[] hardWeightsMultiplier, BigDecimal[] softWeightsMultiplier) {
+    @Override
+    public void impactScore(RuleContext kcontext, BigDecimal weightMultiplier) {
+        throw new UnsupportedOperationException("In the rule (" + kcontext.getRule().getName()
+                + "), the scoreHolder class (" + getClass()
+                + ") does not support a BigDecimal weightMultiplier (" + weightMultiplier + ").\n"
+                + "If you're using constraint streams, maybe switch from penalizeBigDecimal() to penalizeLong().");
+    }
+
+    private void impactScore(RuleContext kcontext, long[] hardWeightsMultiplier, long[] softWeightsMultiplier) {
         Rule rule = kcontext.getRule();
-        ScoreMatchExecutor<BendableBigDecimalScore> matchExecutor = matchExecutorByScoreMap.get(rule);
+        ScoreMatchExecutor<BendableLongScore> matchExecutor = matchExecutorByScoreMap.get(rule);
         if (matchExecutor == null) {
             throw new IllegalStateException("The DRL rule (" + rule.getPackageName() + ":" + rule.getName()
                     + ") does not match a @" + ConstraintWeight.class.getSimpleName() + " on the @"
                     + ConstraintConfiguration.class.getSimpleName() + " annotated class.");
         }
-        matchExecutor.accept(kcontext, BendableBigDecimalScore.of(hardWeightsMultiplier, softWeightsMultiplier));
+        matchExecutor.accept(kcontext, BendableLongScore.of(hardWeightsMultiplier, softWeightsMultiplier));
     }
 
     // ************************************************************************
@@ -221,54 +222,50 @@ public final class BendableBigDecimalScoreHolderImpl extends AbstractScoreHolder
     // ************************************************************************
 
     @Override
-    public void addHardConstraintMatch(RuleContext kcontext, int hardLevel, BigDecimal weight) {
+    public void addHardConstraintMatch(RuleContext kcontext, int hardLevel, long weight) {
         if (hardLevel >= hardScores.length) {
             throw new IllegalArgumentException("The hardLevel (" + hardLevel
                     + ") isn't lower than the hardScores length (" + hardScores.length
                     + ") defined by the @" + PlanningScore.class.getSimpleName() + " on the planning solution class.");
         }
-        hardScores[hardLevel] = hardScores[hardLevel].add(weight);
+        hardScores[hardLevel] += weight;
         registerConstraintMatch(kcontext,
-                () -> hardScores[hardLevel] = hardScores[hardLevel].subtract(weight),
+                () -> hardScores[hardLevel] -= weight,
                 () -> {
-                    BigDecimal[] newHardScores = new BigDecimal[hardScores.length];
-                    Arrays.fill(newHardScores, BigDecimal.ZERO);
-                    BigDecimal[] newSoftScores = new BigDecimal[softScores.length];
-                    Arrays.fill(newSoftScores, BigDecimal.ZERO);
+                    long[] newHardScores = new long[hardScores.length];
+                    long[] newSoftScores = new long[softScores.length];
                     newHardScores[hardLevel] = weight;
-                    return BendableBigDecimalScore.of(newHardScores, newSoftScores);
+                    return BendableLongScore.of(newHardScores, newSoftScores);
                 });
     }
 
     @Override
-    public void addSoftConstraintMatch(RuleContext kcontext, int softLevel, BigDecimal weight) {
+    public void addSoftConstraintMatch(RuleContext kcontext, int softLevel, long weight) {
         if (softLevel >= softScores.length) {
             throw new IllegalArgumentException("The softLevel (" + softLevel
                     + ") isn't lower than the softScores length (" + softScores.length
                     + ") defined by the @" + PlanningScore.class.getSimpleName() + " on the planning solution class.");
         }
-        softScores[softLevel] = softScores[softLevel].add(weight);
+        softScores[softLevel] += weight;
         registerConstraintMatch(kcontext,
-                () -> softScores[softLevel] = softScores[softLevel].subtract(weight),
+                () -> softScores[softLevel] -= weight,
                 () -> {
-                    BigDecimal[] newHardScores = new BigDecimal[hardScores.length];
-                    Arrays.fill(newHardScores, BigDecimal.ZERO);
-                    BigDecimal[] newSoftScores = new BigDecimal[softScores.length];
-                    Arrays.fill(newSoftScores, BigDecimal.ZERO);
+                    long[] newHardScores = new long[hardScores.length];
+                    long[] newSoftScores = new long[softScores.length];
                     newSoftScores[softLevel] = weight;
-                    return BendableBigDecimalScore.of(newHardScores, newSoftScores);
+                    return BendableLongScore.of(newHardScores, newSoftScores);
                 });
     }
 
     @Override
-    public void addMultiConstraintMatch(RuleContext kcontext, BigDecimal[] hardWeights, BigDecimal[] softWeights) {
+    public void addMultiConstraintMatch(RuleContext kcontext, long[] hardWeights, long[] softWeights) {
         if (hardWeights.length != hardScores.length) {
             throw new IllegalArgumentException("The hardWeights length (" + hardWeights.length
                     + ") is different than the hardScores length (" + hardScores.length
                     + ") defined by the @" + PlanningScore.class.getSimpleName() + " on the planning solution class.");
         }
         for (int i = 0; i < hardScores.length; i++) {
-            hardScores[i] = hardScores[i].add(hardWeights[i]);
+            hardScores[i] += hardWeights[i];
         }
         if (softWeights.length != softScores.length) {
             throw new IllegalArgumentException("The softWeights length (" + softWeights.length
@@ -276,23 +273,23 @@ public final class BendableBigDecimalScoreHolderImpl extends AbstractScoreHolder
                     + ") defined by the @" + PlanningScore.class.getSimpleName() + " on the planning solution class.");
         }
         for (int i = 0; i < softScores.length; i++) {
-            softScores[i] = softScores[i].add(softWeights[i]);
+            softScores[i] += softWeights[i];
         }
         registerConstraintMatch(kcontext,
                 () -> {
                     for (int i = 0; i < hardScores.length; i++) {
-                        hardScores[i] = hardScores[i].subtract(hardWeights[i]);
+                        hardScores[i] -= hardWeights[i];
                     }
                     for (int i = 0; i < softScores.length; i++) {
-                        softScores[i] = softScores[i].subtract(softWeights[i]);
+                        softScores[i] -= softWeights[i];
                     }
                 },
-                () -> BendableBigDecimalScore.of(hardWeights, softWeights));
+                () -> BendableLongScore.of(hardWeights, softWeights));
     }
 
     @Override
-    public BendableBigDecimalScore extractScore(int initScore) {
-        return BendableBigDecimalScore.ofUninitialized(initScore,
+    public BendableLongScore extractScore(int initScore) {
+        return BendableLongScore.ofUninitialized(initScore,
                 Arrays.copyOf(hardScores, hardScores.length),
                 Arrays.copyOf(softScores, softScores.length));
     }
