+++ optaplanner-core/src/test/java/org/optaplanner/core/impl/score/director/drools/holder/AbstractScoreHolderTest.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.optaplanner.core.impl.score.buildin;
+package org.optaplanner.core.impl.score.director.drools.holder;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatIllegalStateException;
