+++ optaplanner-core/src/test/java/org/optaplanner/core/impl/score/director/drools/holder/BendableLongScoreHolderImplTest.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.optaplanner.core.impl.score.buildin.bendablelong;
+package org.optaplanner.core.impl.score.director.drools.holder;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatIllegalArgumentException;
@@ -24,7 +24,6 @@ import org.kie.api.definition.rule.Rule;
 import org.kie.api.runtime.rule.RuleContext;
 import org.optaplanner.core.api.score.buildin.bendablelong.BendableLongScore;
 import org.optaplanner.core.api.score.buildin.bendablelong.BendableLongScoreHolder;
-import org.optaplanner.core.impl.score.buildin.AbstractScoreHolderTest;
 
 public class BendableLongScoreHolderImplTest extends AbstractScoreHolderTest<BendableLongScore> {
