+++ optaplanner-core/src/test/java/org/optaplanner/core/impl/score/director/drools/holder/BendableScoreHolderImplTest.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.optaplanner.core.impl.score.buildin.bendable;
+package org.optaplanner.core.impl.score.director.drools.holder;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatIllegalArgumentException;
@@ -24,7 +24,6 @@ import org.kie.api.definition.rule.Rule;
 import org.kie.api.runtime.rule.RuleContext;
 import org.optaplanner.core.api.score.buildin.bendable.BendableScore;
 import org.optaplanner.core.api.score.buildin.bendable.BendableScoreHolder;
-import org.optaplanner.core.impl.score.buildin.AbstractScoreHolderTest;
 
 public class BendableScoreHolderImplTest extends AbstractScoreHolderTest {
