+++ optaplanner-core/src/test/java/org/optaplanner/core/impl/score/director/drools/holder/HardSoftLongScoreHolderImplTest.java
@@ -14,17 +14,16 @@
  * limitations under the License.
  */
 
-package org.optaplanner.core.impl.score.buildin.hardsoft;
+package org.optaplanner.core.impl.score.director.drools.holder;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 import org.junit.jupiter.api.Test;
 import org.kie.api.definition.rule.Rule;
 import org.kie.api.runtime.rule.RuleContext;
-import org.optaplanner.core.api.score.buildin.hardsoft.HardSoftScore;
-import org.optaplanner.core.impl.score.buildin.AbstractScoreHolderTest;
+import org.optaplanner.core.api.score.buildin.hardsoftlong.HardSoftLongScore;
 
-public class HardSoftScoreHolderImplTest extends AbstractScoreHolderTest<HardSoftScore> {
+public class HardSoftLongScoreHolderImplTest extends AbstractScoreHolderTest<HardSoftLongScore> {
 
     @Test
     public void addConstraintMatchWithConstraintMatch() {
@@ -37,48 +36,48 @@ public class HardSoftScoreHolderImplTest extends AbstractScoreHolderTest<HardSof
     }
 
     public void addConstraintMatch(boolean constraintMatchEnabled) {
-        HardSoftScoreHolderImpl scoreHolder = new HardSoftScoreHolderImpl(constraintMatchEnabled);
+        HardSoftLongScoreHolderImpl scoreHolder = new HardSoftLongScoreHolderImpl(constraintMatchEnabled);
 
         RuleContext hard1 = mockRuleContext("hard1");
-        scoreHolder.addHardConstraintMatch(hard1, -1);
-        assertThat(scoreHolder.extractScore(0)).isEqualTo(HardSoftScore.of(-1, 0));
+        scoreHolder.addHardConstraintMatch(hard1, -1L);
+        assertThat(scoreHolder.extractScore(0)).isEqualTo(HardSoftLongScore.of(-1L, -0L));
 
         RuleContext hard2Undo = mockRuleContext("hard2Undo");
-        scoreHolder.addHardConstraintMatch(hard2Undo, -8);
-        assertThat(scoreHolder.extractScore(0)).isEqualTo(HardSoftScore.of(-9, 0));
+        scoreHolder.addHardConstraintMatch(hard2Undo, -8L);
+        assertThat(scoreHolder.extractScore(0)).isEqualTo(HardSoftLongScore.of(-9L, -0L));
         callOnDelete(hard2Undo);
-        assertThat(scoreHolder.extractScore(0)).isEqualTo(HardSoftScore.of(-1, 0));
+        assertThat(scoreHolder.extractScore(0)).isEqualTo(HardSoftLongScore.of(-1L, -0L));
 
         RuleContext soft1 = mockRuleContext("soft1", DEFAULT_JUSTIFICATION, OTHER_JUSTIFICATION);
-        scoreHolder.addSoftConstraintMatch(soft1, -10);
+        scoreHolder.addSoftConstraintMatch(soft1, -10L);
         callOnUpdate(soft1);
-        scoreHolder.addSoftConstraintMatch(soft1, -20); // Overwrite existing
+        scoreHolder.addSoftConstraintMatch(soft1, -20L); // Overwrite existing
 
         RuleContext multi1 = mockRuleContext("multi1");
-        scoreHolder.addMultiConstraintMatch(multi1, -100, -1000);
+        scoreHolder.addMultiConstraintMatch(multi1, -100L, -1000L);
         callOnUpdate(multi1);
-        scoreHolder.addMultiConstraintMatch(multi1, -300, -4000); // Overwrite existing
+        scoreHolder.addMultiConstraintMatch(multi1, -300L, -4000L); // Overwrite existing
 
         RuleContext hard3 = mockRuleContext("hard3");
-        scoreHolder.addHardConstraintMatch(hard3, -10000);
+        scoreHolder.addHardConstraintMatch(hard3, -10000L);
         callOnUpdate(hard3);
-        scoreHolder.addHardConstraintMatch(hard3, -50000); // Overwrite existing
+        scoreHolder.addHardConstraintMatch(hard3, -50000L); // Overwrite existing
 
         RuleContext soft2Undo = mockRuleContext("soft2Undo", UNDO_JUSTIFICATION);
-        scoreHolder.addSoftConstraintMatch(soft2Undo, -99);
+        scoreHolder.addSoftConstraintMatch(soft2Undo, -99L);
         callOnDelete(soft2Undo);
 
         RuleContext multi2Undo = mockRuleContext("multi2Undo");
-        scoreHolder.addMultiConstraintMatch(multi2Undo, -999, -999);
+        scoreHolder.addMultiConstraintMatch(multi2Undo, -999L, -999L);
         callOnDelete(multi2Undo);
 
-        assertThat(scoreHolder.extractScore(0)).isEqualTo(HardSoftScore.of(-50301, -4020));
-        assertThat(scoreHolder.extractScore(-7)).isEqualTo(HardSoftScore.ofUninitialized(-7, -50301, -4020));
+        assertThat(scoreHolder.extractScore(0)).isEqualTo(HardSoftLongScore.of(-50301L, -4020L));
+        assertThat(scoreHolder.extractScore(-7)).isEqualTo(HardSoftLongScore.ofUninitialized(-7, -50301L, -4020L));
         if (constraintMatchEnabled) {
             assertThat(findConstraintMatchTotal(scoreHolder, "hard1").getScore())
-                    .isEqualTo(HardSoftScore.of(-1, 0));
+                    .isEqualTo(HardSoftLongScore.of(-1L, 0L));
             assertThat(scoreHolder.getIndictmentMap().get(OTHER_JUSTIFICATION).getScore())
-                    .isEqualTo(HardSoftScore.of(0, -20));
+                    .isEqualTo(HardSoftLongScore.of(0L, -20L));
             assertThat(scoreHolder.getIndictmentMap().get(UNDO_JUSTIFICATION)).isNull();
         }
     }
@@ -94,27 +93,27 @@ public class HardSoftScoreHolderImplTest extends AbstractScoreHolderTest<HardSof
     }
 
     public void rewardPenalize(boolean constraintMatchEnabled) {
-        HardSoftScoreHolderImpl scoreHolder = new HardSoftScoreHolderImpl(constraintMatchEnabled);
+        HardSoftLongScoreHolderImpl scoreHolder = new HardSoftLongScoreHolderImpl(constraintMatchEnabled);
         Rule hard1 = mockRule("hard1");
-        scoreHolder.configureConstraintWeight(hard1, HardSoftScore.ofHard(10));
+        scoreHolder.configureConstraintWeight(hard1, HardSoftLongScore.ofHard(10L));
         Rule hard2 = mockRule("hard2");
-        scoreHolder.configureConstraintWeight(hard2, HardSoftScore.ofHard(100));
+        scoreHolder.configureConstraintWeight(hard2, HardSoftLongScore.ofHard(100L));
         Rule soft1 = mockRule("soft1");
-        scoreHolder.configureConstraintWeight(soft1, HardSoftScore.ofSoft(10));
+        scoreHolder.configureConstraintWeight(soft1, HardSoftLongScore.ofSoft(10L));
         Rule soft2 = mockRule("soft2");
-        scoreHolder.configureConstraintWeight(soft2, HardSoftScore.ofSoft(100));
+        scoreHolder.configureConstraintWeight(soft2, HardSoftLongScore.ofSoft(100L));
 
         scoreHolder.penalize(mockRuleContext(hard1));
-        assertThat(scoreHolder.extractScore(0)).isEqualTo(HardSoftScore.of(-10, 0));
+        assertThat(scoreHolder.extractScore(0)).isEqualTo(HardSoftLongScore.of(-10L, 0L));
 
-        scoreHolder.penalize(mockRuleContext(hard2), 2);
-        assertThat(scoreHolder.extractScore(0)).isEqualTo(HardSoftScore.of(-210, 0));
+        scoreHolder.penalize(mockRuleContext(hard2), 2L);
+        assertThat(scoreHolder.extractScore(0)).isEqualTo(HardSoftLongScore.of(-210L, 0L));
 
         scoreHolder.reward(mockRuleContext(soft1));
-        assertThat(scoreHolder.extractScore(0)).isEqualTo(HardSoftScore.of(-210, 10));
+        assertThat(scoreHolder.extractScore(0)).isEqualTo(HardSoftLongScore.of(-210L, 10L));
 
-        scoreHolder.reward(mockRuleContext(soft2), 3);
-        assertThat(scoreHolder.extractScore(0)).isEqualTo(HardSoftScore.of(-210, 310));
+        scoreHolder.reward(mockRuleContext(soft2), 3L);
+        assertThat(scoreHolder.extractScore(0)).isEqualTo(HardSoftLongScore.of(-210L, 310L));
     }
 
 }
