+++ optaplanner-core/src/test/java/org/optaplanner/core/impl/score/director/drools/holder/SimpleScoreHolderImplTest.java
@@ -14,17 +14,16 @@
  * limitations under the License.
  */
 
-package org.optaplanner.core.impl.score.buildin.simplelong;
+package org.optaplanner.core.impl.score.director.drools.holder;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 import org.junit.jupiter.api.Test;
 import org.kie.api.definition.rule.Rule;
 import org.kie.api.runtime.rule.RuleContext;
-import org.optaplanner.core.api.score.buildin.simplelong.SimpleLongScore;
-import org.optaplanner.core.impl.score.buildin.AbstractScoreHolderTest;
+import org.optaplanner.core.api.score.buildin.simple.SimpleScore;
 
-public class SimpleLongScoreHolderImplTest extends AbstractScoreHolderTest<SimpleLongScore> {
+public class SimpleScoreHolderImplTest extends AbstractScoreHolderTest<SimpleScore> {
 
     @Test
     public void addConstraintMatchWithConstraintMatch() {
@@ -37,25 +36,25 @@ public class SimpleLongScoreHolderImplTest extends AbstractScoreHolderTest<Simpl
     }
 
     public void addConstraintMatch(boolean constraintMatchEnabled) {
-        SimpleLongScoreHolderImpl scoreHolder = new SimpleLongScoreHolderImpl(constraintMatchEnabled);
+        SimpleScoreHolderImpl scoreHolder = new SimpleScoreHolderImpl(constraintMatchEnabled);
 
         RuleContext scoreRule1 = mockRuleContext("scoreRule1");
-        scoreHolder.addConstraintMatch(scoreRule1, -1000L);
+        scoreHolder.addConstraintMatch(scoreRule1, -1000);
 
         RuleContext scoreRule2 = mockRuleContext("scoreRule2");
-        scoreHolder.addConstraintMatch(scoreRule2, -200L);
+        scoreHolder.addConstraintMatch(scoreRule2, -200);
         callOnDelete(scoreRule2);
 
         RuleContext scoreRule3 = mockRuleContext("scoreRule3");
-        scoreHolder.addConstraintMatch(scoreRule3, -30L);
+        scoreHolder.addConstraintMatch(scoreRule3, -30);
         callOnUpdate(scoreRule3);
-        scoreHolder.addConstraintMatch(scoreRule3, -3L); // Overwrite existing
+        scoreHolder.addConstraintMatch(scoreRule3, -3); // Overwrite existing
 
-        assertThat(scoreHolder.extractScore(0)).isEqualTo(SimpleLongScore.ofUninitialized(0, -1003L));
-        assertThat(scoreHolder.extractScore(-7)).isEqualTo(SimpleLongScore.ofUninitialized(-7, -1003L));
+        assertThat(scoreHolder.extractScore(0)).isEqualTo(SimpleScore.ofUninitialized(0, -1003));
+        assertThat(scoreHolder.extractScore(-7)).isEqualTo(SimpleScore.ofUninitialized(-7, -1003));
         if (constraintMatchEnabled) {
             assertThat(findConstraintMatchTotal(scoreHolder, "scoreRule1").getScore())
-                    .isEqualTo(SimpleLongScore.of(-1000L));
+                    .isEqualTo(SimpleScore.of(-1000));
         }
     }
 
@@ -70,29 +69,29 @@ public class SimpleLongScoreHolderImplTest extends AbstractScoreHolderTest<Simpl
     }
 
     public void rewardPenalize(boolean constraintMatchEnabled) {
-        SimpleLongScoreHolderImpl scoreHolder = new SimpleLongScoreHolderImpl(constraintMatchEnabled);
+        SimpleScoreHolderImpl scoreHolder = new SimpleScoreHolderImpl(constraintMatchEnabled);
         Rule constraint1 = mockRule("constraint1");
-        scoreHolder.configureConstraintWeight(constraint1, SimpleLongScore.of(10L));
+        scoreHolder.configureConstraintWeight(constraint1, SimpleScore.of(10));
         Rule constraint2 = mockRule("constraint2");
-        scoreHolder.configureConstraintWeight(constraint2, SimpleLongScore.of(100L));
+        scoreHolder.configureConstraintWeight(constraint2, SimpleScore.of(100));
 
         scoreHolder.penalize(mockRuleContext(constraint1));
-        assertThat(scoreHolder.extractScore(0)).isEqualTo(SimpleLongScore.of(-10L));
+        assertThat(scoreHolder.extractScore(0)).isEqualTo(SimpleScore.of(-10));
 
-        scoreHolder.penalize(mockRuleContext(constraint2), 2L);
-        assertThat(scoreHolder.extractScore(0)).isEqualTo(SimpleLongScore.of(-210L));
+        scoreHolder.penalize(mockRuleContext(constraint2), 2);
+        assertThat(scoreHolder.extractScore(0)).isEqualTo(SimpleScore.of(-210));
 
-        scoreHolder = new SimpleLongScoreHolderImpl(constraintMatchEnabled);
+        scoreHolder = new SimpleScoreHolderImpl(constraintMatchEnabled);
         Rule constraint3 = mockRule("constraint3");
-        scoreHolder.configureConstraintWeight(constraint3, SimpleLongScore.of(10L));
+        scoreHolder.configureConstraintWeight(constraint3, SimpleScore.of(10));
         Rule constraint4 = mockRule("constraint4");
-        scoreHolder.configureConstraintWeight(constraint4, SimpleLongScore.of(100L));
+        scoreHolder.configureConstraintWeight(constraint4, SimpleScore.of(100));
 
         scoreHolder.reward(mockRuleContext(constraint3));
-        assertThat(scoreHolder.extractScore(0)).isEqualTo(SimpleLongScore.of(10L));
+        assertThat(scoreHolder.extractScore(0)).isEqualTo(SimpleScore.of(10));
 
-        scoreHolder.reward(mockRuleContext(constraint4), 3L);
-        assertThat(scoreHolder.extractScore(0)).isEqualTo(SimpleLongScore.of(310L));
+        scoreHolder.reward(mockRuleContext(constraint4), 3);
+        assertThat(scoreHolder.extractScore(0)).isEqualTo(SimpleScore.of(310));
     }
 
 }
